Fix/forgot password

+++ auth/auth/src/main/java/com/unihub/auth/security/dto/response/AccessToken.java
@@ -15,5 +15,5 @@ public class AccessToken {
     private String tokenType;
 
     @JsonProperty("expires_in")
-    private int expiresIn;
+    private long expiresIn;
 }
